requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
six
pillow
python-decouple
channels
pydantic
channels_redis
daphne
psycopg2